cli/pyproject.toml
[project]
name = "planoai"
version = "0.4.0"
description = "Python-based CLI tool to manage Plano."
authors = [{name = "Katanemo Labs, Inc."}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7,<9.0.0",
    "jinja2>=3.1.4,<4.0.0",
    "jsonschema>=4.23.0,<5.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "requests>=2.31.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1,<9.0.0",
]

[project.scripts]
planoai = "planoai.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "planoai/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["planoai"]

[tool.pytest.ini_options]
addopts = ["-v"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
